Added: hidden content box
The javascript has been simplified for the section containing the roadmap tasks. the tasks have a functionality that when they are clicked on they display the contents of the task. This is now handled with a HiddenContentBox script that automatically searches for elements with the right classes and applies the hidden content functionality. This meant some code could be removed that previously achieved this with ids and arrays.